+++ aioxmpp/muc/service.py
@@ -36,6 +36,41 @@ from . import xso as muc_xso
 
 
 class LeaveMode(Enum):
+    """
+    The different reasons for a user to leave or be removed from MUC.
+
+    .. attribute:: DISCONNECTED
+
+       The local client disconnected. This only occurs in events referring to
+       the local entity.
+
+    .. attribute:: SYSTEM_SHUTDOWN
+
+       The remote server shut down.
+
+    .. attribute:: NORMAL
+
+       The leave was initiated by the occupant themselves and was not a kick or
+       ban.
+
+    .. attribute:: KICKED
+
+       The user was kicked from the room.
+
+    .. attribute:: AFFILIATION_CHANGE
+
+       Changes in the affiliation of the user caused them to be removed.
+
+    .. attribute:: MODERATION_CHANGE
+
+       Changes in the moderation settings of the room caused the user to be
+       removed.
+
+    .. attribute:: BANNED
+
+       The user was banned from the room.
+    """
+
     DISCONNECTED = -2
     SYSTEM_SHUTDOWN = -1
     NORMAL = 0
@@ -642,7 +677,9 @@ class Room:
                                    self._mucjid,
                                    stanza)
 
-        if 110 in stanza.xep0045_muc_user.status_codes:
+        if (110 in stanza.xep0045_muc_user.status_codes or
+                (self._this_occupant is not None and
+                 self._this_occupant.occupantjid == stanza.from_)):
             self._service.logger.debug("%s: is self-presence",
                                        self._mucjid)
             self._handle_self_presence(stanza)
@@ -652,6 +689,13 @@ class Room:
         try:
             existing = self._occupant_info[info.occupantjid]
         except KeyError:
+            if stanza.type_ == aioxmpp.structs.PresenceType.UNAVAILABLE:
+                self._service.logger.debug(
+                    "received unavailable presence from unknown occupant %r."
+                    " ignoring.",
+                    stanza.from_,
+                )
+                return
             self._occupant_info[info.occupantjid] = info
             self.on_join(stanza, info)
             return
@@ -769,7 +813,7 @@ class Room:
 
         self.leave()
 
-        def on_exit(*args):
+        def on_exit(*args, **kwargs):
             fut.set_result(None)
             return True