+++ docs/api/changelog.rst
@@ -53,6 +53,16 @@ Version 0.8
   * :meth:`~.AbstractClient.on_stream_destroyed` got a new argument `reason`
     which gives the exception which caused the stream to be destroyed.
 
+* Improve compatibility of :mod:`aioxmpp.muc` with Prosody 0.9 and below, which
+  misses sending the ``110`` status code on some presences.
+
+* Fix :meth:`aioxmpp.muc.Room.leave_and_wait` never returning.
+
+* Do not emit :meth:`aioxmpp.muc.Room.on_join` when an unavailable presence from
+  an unknown occupant JID is received.
+
+* :mod:`aioxmpp.e2etest`: Framework for End-to-End/Integration tests.
+
 .. _api-changelog-0.7:
 
 Version 0.7